+++ Mod Projects/Lumiore/Buff/B_Lumiore_13Rare.cs
@@ -55,7 +55,7 @@ namespace Lumiore
 
         public override string DescExtended()
         {
-            return base.DescExtended().Replace("%a", (this.BChar.GetStat.atk * 2.0f).ToString());
+            return base.DescExtended().Replace("%a", ((int)(this.BChar.GetStat.atk * 2.0f)).ToString());
         }
     }
 }
